fix(core): use float64 instead of float32 (#4)

+++ average.go
@@ -1,14 +1,14 @@
 package stats
 
 // Average gets the number expressing the central or typical value in a set of data
-func Average(datas []float32) float32 {
+func Average(datas []float64) float64 {
 	length := len(datas)
 	if length == 0 {
 		return 0
 	}
-	ret := float32(0)
+	ret := float64(0)
 	for i := 0; i < length; i++ {
 		ret += datas[i]
 	}
-	return ret / float32(length)
+	return ret / float64(length)
 }
